pyproject.toml
[project]
name = "autonomous-racing"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
     "gpxpy>=1.5.0",
    "matplotlib>=3.5.0",
    "numpy>=1.21.0,<2.0",
    "pyyaml>=6.0",
    "scipy>=1.8.1",
    "shapely>=2.0.1",
    "rerun-sdk>=0.20.0",
]
